replace stream module by improved version of readable/writable impl
